+++ app/Http/Controllers/Api/V1/WalletController.php
@@ -8,6 +8,9 @@ use App\Http\Resources\V1\WalletResource;
 use App\Http\Resources\V1\WalletCollection;
 use App\Http\Requests\V1\StoreWalletRequest;
 use App\Http\Requests\V1\UpdateWalletRequest;
+use Illuminate\Support\Facades\Auth;
+use Symfony\Component\HttpFoundation\Response;
+
 
 class WalletController extends Controller
 {
@@ -39,7 +42,19 @@ class WalletController extends Controller
      */
     public function store(StoreWalletRequest $request)
     {
-        return new WalletResource(Wallet::create($request->all()));
+        if(auth()->user()->role == "admin")
+        {
+            return new WalletResource(Wallet::create($request->all()));
+            return response([
+                'message' => 'Success'
+            ])->withCookie($cookie);
+        }
+        else
+        {
+            return response([
+                'message' => 'You are unauthorized!'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -50,7 +65,19 @@ class WalletController extends Controller
      */
     public function show(Wallet $wallet)
     {
-        return new WalletResource($wallet);
+        if(auth()->user()->role == "admin" || auth()->user()->role == "user")
+        {
+            return new WalletResource($wallet);
+            return response([
+                'message' => 'Success'
+            ]);
+        }
+        else
+        {
+            return response([
+                'message' => 'You are unauthorized!'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -73,7 +100,19 @@ class WalletController extends Controller
      */
     public function update(UpdateWalletRequest $request, Wallet $wallet)
     {
-        $wallet ->update($request->all());
+        if(auth()->user()->role == "admin")
+        {
+            $wallet ->update($request->all());
+            return response([
+                'message' => 'Success'
+            ])->withCookie($cookie);
+        }
+        else
+        {
+            return response([
+                'message' => 'You are unauthorized!'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 
     /**
@@ -84,7 +123,19 @@ class WalletController extends Controller
      */
     public function destroy(Wallet $wallet)
     {
-        $wallet->delete();
-        return response(null, 204);
+        if(auth()->user()->role == "admin")
+        {
+            $wallet->delete();
+            return response(null, 204);
+            return response([
+                'message' => 'Success'
+            ])->withCookie($cookie);
+        }
+        else
+        {
+            return response([
+                'message' => 'You are unauthorized!'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
     }
 }
